Refactored AngleIterator to be AngleCalculator
