+++ shortcut_composer/templates/pie_menu_utils/__init__.py
@@ -1,4 +1,4 @@
-from .angle_iterator import AngleIterator
+from .angle_calculator import AngleCalculator
 from .label_holder import LabelHolder
 from .pie_manager import PieManager
 from .pie_widget import PieWidget
@@ -6,7 +6,7 @@ from .pie_style import PieStyle
 from .label import Label
 
 __all__ = [
-    "AngleIterator",
+    "AngleCalculator",
     "LabelHolder",
     "PieManager",
     "PieWidget",
